feat: 상품 목록 조회에 판매 최저가 추가
-상품 목록 조회에 판매 최저가 추

issue number: #77

+++ src/main/java/com/flab/shoeauction/controller/dto/ProductDto.java
@@ -199,7 +199,7 @@ public class ProductDto {
         private String brandThumbnailImagePath;
         private String nameKor;
         private String nameEng;
-        private int lowestPrice;
+        private Long lowestPrice;
     }
 
     @Getter
